requirements.txt
flask
pandas
splitwise
python-dotenv
